+++ simplexss/core/enums/graphic.py
@@ -2,5 +2,4 @@ from enum import Enum
 
 
 class GraphicMode(str, Enum):
-    CLI = 'cli'
     GUI = 'gui'