+++ simplexss/core/enums/__init__.py
@@ -1,0 +1,5 @@
+from simplexss.core.enums.graphics import GraphicMode
+
+__all__ = [
+    "GraphicMode",
+]