+++ simplexss/core/enums/graphics.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class GraphicMode(str, Enum):
+    CONSOLE = "console"
+    BROWSER = "browser"
+    DESKTOP = "desktop"